Make it easier to switch between lua mains
